backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment & Config
python-decouple==3.8

# Authentication (JWT)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Forms/File uploads
python-multipart==0.0.6

# CORS (for Vue frontend)
fastapi==0.104.1  # includes CORS middleware